# docs/swagger/swagger.yaml
basePath: /
...
definitions:
...
  http.ErrorResponse:
    description: Standard error response returned by all API endpoints
    properties:
      code:
        description: HTTP status code
        example: 400
        type: integer
      message:
        description: Human-readable error message
        example: context name is required
        type: string
      title:
        description: Error type identifier
        example: invalid_request
        type: string
    type: object
...
  internal_matching_adapters_http.UnmatchRequest:
    properties:
      reason:
        example: incorrect match - amounts do not match
        type: string
    required:
    - reason
    type: object
...
info:
  contact:
    email: support@lerian.studio
    name: Lerian Studio Support
    url: https://discord.gg/DnhqKwkGv3
  description: |-
    Reconciliation engine for the Lerian Studio ecosystem.
    Provides automated transaction matching between Midaz ledger and external systems.
  license:
    name: Lerian Studio General License
    url: https://github.com/LerianStudio/matcher/blob/main/LICENSE.md
  termsOfService: http://swagger.io/terms/
  title: Matcher Reconciliation API
  version: v1.0.0
paths:
...
  /v1/matching/groups/{matchGroupId}:
    delete:
      consumes:
      - application/json
      description: Breaks an incorrect match group, rejecting it with a reason and
        reverting all associated transactions to UNMATCHED status.
      operationId: unmatch
      parameters:
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Match Group ID
        format: uuid
        in: path
        name: matchGroupId
        required: true
        type: string
      - description: Context ID
        format: uuid
        in: query
        name: contextId
        required: true
        type: string
      - description: Unmatch payload with rejection reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_matching_adapters_http.UnmatchRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Match group not found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Break/Unmatch a match group
      tags:
      - Matching
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication (format: "Bearer {token}")'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"